Added cartService (Working).

+++ src/main/java/com/cleancode/shopping/payload/ProductResponse.java
@@ -15,6 +15,6 @@ public class ProductResponse {
     private String productName;
     //private long productId;
     private long quantity;
-    private long price;
+    private Double price;
     private Blob image;
 }
